+++ database/migrations/-10000_00_00_000002_create_regions_table.php
@@ -14,10 +14,10 @@ class CreateRegionsTable extends Migration
     public function up()
     {
         Schema::create('regions', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('code', 10);
-            $table->integer('country_id')->unsigned();
+            $table->bigInteger('country_id')->unsigned();
 
             $table->foreign('country_id')
                 ->references('id')
