+++ database/migrations/-10000_00_00_000003_create_cities_table.php
@@ -14,9 +14,9 @@ class CreateCitiesTable extends Migration
     public function up()
     {
         Schema::create('cities', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('region_id')->unsigned();
-            $table->integer('country_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('region_id')->unsigned();
+            $table->bigInteger('country_id')->unsigned();
             $table->decimal('latitude', 10, 8);
             $table->decimal('longitude', 11, 8);
             $table->string('name');
