+++ database/migrations/0000_00_00_000001_create_users_table.php
@@ -14,16 +14,16 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('auth_users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('email')->unique();
             $table->string('password');
             $table->boolean('disabled')->default(0);
-            $table->integer('role_id')->nullable();
+            $table->bigInteger('role_id')->nullable();
             $table->string('uid')->nullable();
 
-            $table->integer('createdBy')->nullable();
-            $table->integer('updatedBy')->nullable();
+            $table->bigInteger('createdBy')->nullable();
+            $table->bigInteger('updatedBy')->nullable();
             $table->rememberToken();
             $table->timestamps();
             $table->softDeletes();
