+++ database/migrations/0000_00_00_000004_create_users_roles_table.php
@@ -14,17 +14,20 @@ class CreateUsersRolesTable extends Migration
     public function up()
     {
         Schema::create('auth_users_has_roles', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned();
-            $table->integer('role_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
 
             $table->primary(['user_id', 'role_id']);
 
-            $table->foreign('user_id')->references('id')
-                ->on('auth_users')->onDelete('cascade');
+            $table->foreign('user_id')
+                ->references('id')
+                ->on('auth_users')
+                ->onDelete('cascade');
 
-            $table->foreign('role_id')->references('id')
-                ->on('auth_roles')->onDelete('cascade');
-
+            $table->foreign('role_id')
+                ->references('id')
+                ->on('auth_roles')
+                ->onDelete('cascade');
         });
     }
