+++ database/migrations/0000_00_00_000007_create_accounts_table.php
@@ -24,12 +24,12 @@ class CreateAccountsTable extends Migration
             $table->string('address')->nullable();
             $table->string('neighborhood')->nullable();
 
-            $table->integer('country_id')->unsigned()->nullable();
-            $table->integer('region_id')->unsigned()->nullable();
-            $table->integer('city_id')->unsigned()->nullable();
+            $table->bigInteger('country_id')->unsigned()->nullable();
+            $table->bigInteger('region_id')->unsigned()->nullable();
+            $table->bigInteger('city_id')->unsigned()->nullable();
 
-            $table->integer('createdBy')->nullable();
-            $table->integer('updatedBy')->nullable();
+            $table->bigInteger('createdBy')->nullable();
+            $table->bigInteger('updatedBy')->nullable();
             $table->timestamps();
             $table->softDeletes();
